[Fixed] 별점 기능 추가
TotalRate 더블형으로 변수 수정
rateAvg 변수 추가

StoreService 코드 추가
rateAvg 계산 코드 추가

StoreResponseDto ...
getAllStore, getStore ..
rateAvg 인자추가

+++ src/main/java/com/example/Fooding/review/dto/RequestReview.java
@@ -12,7 +12,7 @@ public class RequestReview {
     @Builder
     public static class CreateReviewDto {
         private String reviewContent;
-        private Long rate;
+        private Double rate;
         private Long storeId;
 
         public static Review toEntity(CreateReviewDto createReviewDto, Store store, Long makerId) {
@@ -33,7 +33,7 @@ public class RequestReview {
     public static class UpdateReviewDto {
         private Long reviewId;
         private String reviewContent;
-        private Long rate;
+        private Double rate;
 
         public static Review toEntity(Review review, UpdateReviewDto updateReviewDto) {
             review.update(updateReviewDto.getReviewContent(), updateReviewDto.getRate());
